[update] 댓글 조회 클라이언트 액션으로 수정
