+++ src/utils/supabase/client-actions.ts
@@ -82,3 +82,47 @@ export const deleteProfileImg = async (user: User) => {
     });
   }
 };
+
+// 댓글 조회
+export async function fetchComment(postId: string) {
+  const supabase = createClient();
+  const STORAGE = "profiles";
+
+  const { data: comments, error: commentError } = await supabase
+    .from("comments")
+    .select("comment_id, content, user_id, created_at")
+    .eq("post_id", postId);
+
+  if (commentError) {
+    throw new Error("댓글을 불러오는데 실패했습니다.");
+  }
+
+  const commentsWithProfile = await Promise.all(
+    comments.map(async (comment) => {
+      const { data: profile, error: profileError } = await supabase
+        .from("profiles")
+        .select("nickname, profile_img")
+        .eq("user_id", comment.user_id)
+        .single();
+
+      if (profileError) {
+        throw new Error("프로필 정보를 불러오는데 실패했습니다.");
+      }
+
+      // `profile_img`를 가져와 절대 경로 생성
+      const { data: { publicUrl: profileImgUrl } = {} } = supabase.storage
+        .from(STORAGE)
+        .getPublicUrl(profile?.profile_img ?? "default");
+
+      return {
+        ...comment,
+        profile: {
+          nickname: profile?.nickname,
+          profile_img: profileImgUrl || "/default-profile.png"
+        }
+      };
+    })
+  );
+
+  return commentsWithProfile;
+}
